cleanup, move utils into hook
